Adding authentication to all routes

+++ server/server.js
@@ -25,10 +25,11 @@ const port = process.env.PORT;
 
 app.use(bodyParser.json());
 
-app.post('/todos',(req,res) => {
+app.post('/todos', authenticate, (req,res) => {
     console.log(req.body);
     var todo = new Todo({
-      text:req.body.text
+      text:req.body.text,
+      _creator:req.user._id
     });
 
     todo.save().then((resp) => {
@@ -38,21 +39,23 @@ app.post('/todos',(req,res) => {
     })
 });
 
-app.get('/todos',(req,res) => {
-  Todo.find().then((todos) => {
+app.get('/todos', authenticate, (req,res) => {
+  Todo.find({_creator:req.user._id}).then((todos) => {
     res.send({todos});
   }).catch((err) => {
     res.status(400).send(err);
   })
 });
 
-app.get('/todos/:id',(req,res) => {
+app.get('/todos/:id', authenticate, (req,res) => {
   var id = req.params.id;
+  console.log(id);
+  console.log(req.user._id);
   if (!ObjectID.isValid(id)){
     return res.status(404).send('Not a valid ID');
   }
   else {
-    Todo.findById(id).then((resp) => {
+    Todo.findOne({_id:id,_creator:req.user._id}).then((resp) => {
       if (!resp) {
         return res.status(404).send();
       }
@@ -63,13 +66,13 @@ app.get('/todos/:id',(req,res) => {
   }
 });
 
-app.delete('/todos/:id',(req,res) => {
+app.delete('/todos/:id', authenticate, (req,res) => {
   var id = req.params.id;
   if (!ObjectID.isValid(id)){
     return res.status(404).send('Not a valid ID');
   }
   else {
-    Todo.deleteOne({_id:new ObjectID(id)}).then((resp) => {
+    Todo.deleteOne({_id:new ObjectID(id),_creator:req.user._id}).then((resp) => {
       if (!resp) {
         return res.status(404).send();
       }
@@ -80,7 +83,7 @@ app.delete('/todos/:id',(req,res) => {
   }
 });
 
-app.patch('/todos/:id',(req,res) => {
+app.patch('/todos/:id', authenticate, (req,res) => {
   var id = req.params.id;
   var body = _.pick(req.body,['text','completed']);
 
@@ -95,7 +98,7 @@ app.patch('/todos/:id',(req,res) => {
       body.completed = false;
       body.completedAt = null;
     }
-    Todo.updateOne({_id:id},{$set:body},{new:true}).then((resp) => {
+    Todo.updateOne({_id:id,_creator:req.user._id},{$set:body},{new:true}).then((resp) => {
         if(!resp){
           return res.status(400).send();
         }
@@ -140,7 +143,8 @@ app.post('/users/login',(req,res) => {
 });
 
 app.delete('/users/me/token', authenticate, (req,res) => {
-    req.user.removeToken(req.token).then(() => {
+    req.user.removeToken(req.token).then((test) => {
+      console.log('User object from token delete: '+JSON.stringify(test));
       res.status(200).send();
     },() => {
       res.status(400).send();
